Don't build nil translations in AR table backend
In some cases we might want to edit all translations for a record
simultaneously and allow users to select the locales the record is
translated to.

The code could be something like:

```
I18n.available_locales.each do |locale|
  translation = object.translations.in_locale(locale) ||
    object.translations.new(locale: locale).tap(&:mark_for_destruction)

  fields_for(:translations, translation) do |fields|
    fields.hidden_field :_destroy
    # ...
  end
end
```

This code marks for destruction translations which haven't been selected
by users, and then using a JavaScript interface users could change the
value of the `_destroy` field of each translation.

However, this code wouldn't work with Mobility's fallbacks. In a
situation where we display the title of an article before the form to
edit it, the current locale is English, the article only exists in
French, and English has fallbacks to German, Japanese, Spanish, Italian
and French:

```
<%= @article.title %>
<% # Insert the previous code here %>
```

This code would create Translation objects for English, German,
Japanese, Spanish and Italian, and so all these locales would be shown
as if users had selected them.

Not building translation objects when simply reading attributes solves
the issue.